Add migration to add db table for WorkElement
+ Some minor fixes in naming

+++ RepairsApi/V1/Infrastructure/Quantity.cs
@@ -1,4 +1,3 @@
-using Hackney.Housing.Hact;
 using Microsoft.EntityFrameworkCore;
 
 namespace RepairsApi.V1.Infrastructure
@@ -6,7 +5,7 @@ namespace RepairsApi.V1.Infrastructure
     [Owned]
     public class Quantity
     {
-        public double Amount { get; set; }
+        public int Amount { get; set; }
         public string UnitOfMeasurementCode { get; set; }
     }
 }
